backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
gunicorn==25.1.0
Pillow==12.1.1
